Revert "fix: add field auth on aggregation queries (#8508)"
This reverts commit c0fa85a87230d631ffaf376f18f4fc3c4ec9a1f9.

The commit was causing E2E test failures.
